Separate main and bank2ynab modules
